# Conflicts:
#	android/cli/commands/_build.js
#	android/modules/android/src/java/ti/modules/titanium/android/notificationmanager/NotificationProxy.java
#	android/modules/gesture/src/java/ti/modules/titanium/gesture/GestureModule.java
#	android/modules/network/src/java/ti/modules/titanium/network/TiHTTPClient.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/ButtonProxy.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/LabelProxy.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/NotificationProxy.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/TableViewRowProxy.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/WebViewProxy.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/WindowProxy.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUIActivityIndicator.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUIButton.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUIDialog.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUILabel.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUINotification.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUIScrollableView.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/widget/TiUIText.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/widget/listview/ListViewProxy.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/widget/listview/TiListView.java
#	android/modules/ui/src/java/ti/modules/titanium/ui/widget/webview/TiUIWebView.java
#	android/runtime/common/src/js/module.js
#	android/runtime/common/src/js/titanium.js
#	android/templates/build/theme.xml
#	android/titanium/src/java/org/appcelerator/titanium/TiBaseActivity.java
#	android/titanium/src/java/org/appcelerator/titanium/TiBlob.java
#	android/titanium/src/java/org/appcelerator/titanium/TiC.java
#	android/titanium/src/java/org/appcelerator/titanium/TiFileProxy.java
#	android/titanium/src/java/org/appcelerator/titanium/view/TiUIView.java
#	iphone/Classes/APSHTTPClient
#	iphone/Classes/APSHTTPClient/libAPSHTTPClient.a
#	iphone/Classes/KrollBridge.m
#	iphone/Classes/MediaModule.h
#	iphone/Classes/MediaModule.m
#	iphone/Classes/TiBlob.m
#	iphone/Classes/TiUIButton.m
#	iphone/Classes/TiUILabel.m
#	iphone/Classes/TiUIListItem.m
#	iphone/Classes/TiUITableViewRowProxy.m
#	iphone/Classes/TiUITextField.m
#	iphone/Classes/TiUIWebView.m
#	iphone/Classes/TiUIWindowProxy.m
#	iphone/Classes/TiWindowProxy.m
#	iphone/Classes/TopTiModule.m
#	iphone/cli/commands/_build.js
#	iphone/cli/commands/_buildModule.js
#	iphone/iphone/Titanium.xcodeproj/project.pbxproj

+++ iphone/cli/commands/_build.js
@@ -991,13 +991,13 @@ iOSBuilder.prototype.configOptionPPuuid = function configOptionPPuuid(order) {
 						provisioningProfiles[__('Available Development UUIDs:')] = pp;
 					} else {
 						logger.error(__('Unable to find any non-expired development provisioning profiles that match the app id "%s"', appId) + '\n');
-						logger.log(__('You will need to login into %s with your Apple Download account, then create, download, and install a profile.',
+						logger.log(__('You will need to log in to %s with your Apple Developer account, then create, download, and install a profile.',
 							'http://appcelerator.com/ios-dev-certs'.cyan) + '\n');
 						process.exit(1);
 					}
 				} else {
 					logger.error(__('Unable to find any development provisioning profiles') + '\n');
-					logger.log(__('You will need to login into %s with your Apple Download account, then create, download, and install a profile.',
+					logger.log(__('You will need to log in to %s with your Apple Developer account, then create, download, and install a profile.',
 						'http://appcelerator.com/ios-dev-certs'.cyan) + '\n');
 					process.exit(1);
 				}
@@ -1017,13 +1017,13 @@ iOSBuilder.prototype.configOptionPPuuid = function configOptionPPuuid(order) {
 
 					if (!valid) {
 						logger.error(__('Unable to find any non-expired distribution or adhoc provisioning profiles that match the app id "%s".', appId) + '\n');
-						logger.log(__('You will need to login into %s with your Apple Download account, then create, download, and install a profile.',
+						logger.log(__('You will need to log in to %s with your Apple Developer account, then create, download, and install a profile.',
 							'http://appcelerator.com/ios-dist-certs'.cyan) + '\n');
 						process.exit(1);
 					}
 				} else {
 					logger.error(__('Unable to find any distribution or adhoc provisioning profiles'));
-					logger.log(__('You will need to login into %s with your Apple Download account, then create, download, and install a profile.',
+					logger.log(__('You will need to log in to %s with your Apple Developer account, then create, download, and install a profile.',
 						'http://appcelerator.com/ios-dist-certs'.cyan) + '\n');
 					process.exit(1);
 				}
@@ -1880,14 +1880,14 @@ iOSBuilder.prototype.validate = function (logger, config, cli) {
 			},
 
 			function selectDevice(next) {
-				if (cli.argv.target === 'dist-appstore' || cli.argv.target === 'dist-adhoc' || cli.argv['build-only']) {
+				if (cli.argv.target === 'dist-appstore' || cli.argv.target === 'dist-adhoc') {
 					return next();
 				}
 
 				// no --device-id or doing a build-only sim build, so pick a device
 
 				if (cli.argv.target === 'device') {
-					if (!cli.argv['device-id']) {
+					if (!cli.argv['build-only'] && !cli.argv['device-id']) {
 						cli.argv['device-id'] = this.iosInfo.devices.length ? this.iosInfo.devices[0].udid : 'itunes';
 					}
 					return next();
@@ -2046,24 +2046,51 @@ iOSBuilder.prototype.validate = function (logger, config, cli) {
 						if (module.platform.indexOf('commonjs') !== -1) {
 							module.native = false;
 
-							// Look for legacy module.id.js first
-							module.libFile = path.join(module.modulePath, module.id + '.js');
-							if (!fs.existsSync(module.libFile)) {
-								// then package.json TODO Verify the main property points at reale file under the module!
-								module.libFile = path.join(module.modulePath, 'package.json');
-								if (!fs.existsSync(module.libFile)) {
-									// then index.js
-									module.libFile = path.join(module.modulePath, 'index.js');
-									if (!fs.existsSync(module.libFile)) {
-										// then index.json
-										module.libFile = path.join(module.modulePath, 'index.json');
-										if (!fs.existsSync(module.libFile)) {
-											this.logger.error(__('Module %s version %s is missing module files: %s, package.json, index.js, or index.json', module.id.cyan, (module.manifest.version || 'latest').cyan, path.join(module.modulePath, module.id + '.js').cyan) + '\n');
-											process.exit(1);
+							// look for legacy module.id.js first
+							var libFile = path.join(module.modulePath, module.id + '.js');
+							module.libFile = fs.existsSync(libFile) ? libFile : null;
+							// If no legacy file, look for package.json...
+							if (!module.libFile) {
+								var pkgJsonFile = path.join(module.modulePath, 'package.json');
+								if (fs.existsSync(pkgJsonFile)) {
+									try {
+										var pkgJson = require(pkgJsonFile);
+										// look for 'main' property
+										if (pkgJson && pkgJson.main) {
+											// look for main file as-is
+											if (fs.existsSync(libFile = path.join(module.modulePath, pkgJson.main))) {
+												module.libFile = libFile;
+											}
+											// look with .js extension
+											if (!module.libFile && fs.existsSync(libFile = path.join(module.modulePath, pkgJson.main + '.js'))) {
+												module.libFile = libFile;
+											}
+											// look with .json extension
+											if (!module.libFile && fs.existsSync(libFile = path.join(module.modulePath, pkgJson.main + '.json'))) {
+												module.libFile = libFile;
+											}
 										}
+									} catch (e) {
+										// squeltch
 									}
 								}
+
+								// look for index.js in root directory of module
+								if (!module.libFile && fs.existsSync(libFile = path.join(module.modulePath, 'index.js'))) {
+									module.libFile = libFile;
+								}
+
+								// look for index.json in root directory of module
+								if (!module.libFile && fs.existsSync(libFile = path.join(module.modulePath, 'index.json'))) {
+									module.libFile = libFile;
+								}
+
+								if (!module.libFile) {
+									this.logger.error(__('Module "%s" v%s is missing main file: %s, package.json with "main" entry, index.js, or index.json', module.id, module.manifest.version || 'latest', module.id + '.js') + '\n');
+									process.exit(1);
+								}
 							}
+
 							this.commonJsModules.push(module);
 						} else {
 							module.native = true;
@@ -3822,7 +3849,7 @@ iOSBuilder.prototype.createXcodeProject = function createXcodeProject(next) {
 	hook(xcodeProject, next);
 };
 
-iOSBuilder.prototype._embedCapabilitiesAndWriteEntitlementsPlist = function _embedCapabilitiesAndWriteEntitlementsPlist(plist, dest) {
+iOSBuilder.prototype._embedCapabilitiesAndWriteEntitlementsPlist = function _embedCapabilitiesAndWriteEntitlementsPlist(plist, dest, isExtension, next) {
 	var caps = this.tiapp.ios.capabilities,
 		parent = path.dirname(dest);
 
@@ -3840,25 +3867,42 @@ iOSBuilder.prototype._embedCapabilitiesAndWriteEntitlementsPlist = function _emb
 
 	this.unmarkBuildDirFile(dest);
 
-	// write the entitlements.plist
-	var contents = plist.toString('xml');
-	if (!fs.existsSync(dest) || contents !== fs.readFileSync(dest).toString()) {
-		if (!this.forceRebuild) {
-			this.logger.info(__('Forcing rebuild: %s has changed since last build', dest.replace(this.projectDir + '/', '')));
-			this.forceRebuild = true;
+	var rel = path.relative(this.buildDir, dest);
+	if (['ios', 'iphone'].some(function (dir) {
+		if (fs.existsSync(path.join(this.projectDir, 'platform', dir, rel))) {
+			return true;
 		}
-		this.logger.debug(__('Writing %s', dest.cyan));
-		fs.existsSync(parent) || wrench.mkdirSyncRecursive(parent);
-		fs.writeFileSync(dest, contents);
-	} else {
-		this.logger.trace(__('No change, skipping %s', dest.cyan));
+	}, this)) {
+		return next();
 	}
+
+	var name = 'build.ios.write' + (isExtension ? 'Extension' : '') + 'Entitlements';
+	var hook = this.cli.createHook(name, this, function (plist, dest, done) {
+		// write the entitlements.plist
+		var contents = plist.toString('xml');
+
+		if (!fs.existsSync(dest) || contents !== fs.readFileSync(dest).toString()) {
+			if (!this.forceRebuild) {
+				this.logger.info(__('Forcing rebuild: %s has changed since last build', dest.replace(this.projectDir + '/', '')));
+				this.forceRebuild = true;
+			}
+			this.logger.debug(__('Writing %s', dest.cyan));
+			fs.existsSync(parent) || wrench.mkdirSyncRecursive(parent);
+			fs.writeFileSync(dest, contents);
+		} else {
+			this.logger.trace(__('No change, skipping %s', dest.cyan));
+		}
+
+		done();
+	});
+
+	hook(plist, dest, next);
 };
 
-iOSBuilder.prototype.writeEntitlementsPlist = function writeEntitlementsPlist() {
+iOSBuilder.prototype.writeEntitlementsPlist = function writeEntitlementsPlist(next) {
 	if (!this.forceRebuild || this.cli.argv.xcode) {
 		delete this.buildDirFiles[path.join(this.buildDir, this.tiapp.name + '.entitlements')];
-		return;
+		return next();
 	}
 	this.logger.info(__('Creating Entitlements.plist'));
 
@@ -3909,7 +3953,7 @@ iOSBuilder.prototype.writeEntitlementsPlist = function writeEntitlementsPlist() 
 		}
 	}
 
-	this._embedCapabilitiesAndWriteEntitlementsPlist(plist, path.join(this.buildDir, this.tiapp.name + '.entitlements'));
+	this._embedCapabilitiesAndWriteEntitlementsPlist(plist, path.join(this.buildDir, this.tiapp.name + '.entitlements'), false, next);
 };
 
 iOSBuilder.prototype.writeInfoPlist = function writeInfoPlist() {
@@ -4663,13 +4707,13 @@ iOSBuilder.prototype.copyTitaniumiOSFiles = function copyTitaniumiOSFiles() {
 
 iOSBuilder.prototype.copyExtensionFiles = function copyExtensionFiles() {
 	if (!this.forceRebuild || this.cli.argv.xcode) {
-		return;
+		return next();
 	}
-	if (!this.extensions.length) return;
+	if (!this.extensions.length) return next();
 
 	this.logger.info(__('Copying iOS extensions'));
 
-	this.extensions.forEach(function (extension) {
+	async.eachSeries(this.extensions, function (extension, next) {
 		var extName = path.basename(extension.projectPath).replace(/\.xcodeproj$/, ''),
 			src = path.dirname(extension.projectPath),
 			dest = path.join(this.buildDir, 'extensions', path.basename(src));
@@ -4754,15 +4798,15 @@ iOSBuilder.prototype.copyExtensionFiles = function copyExtensionFiles() {
 		extension.projectPath = path.join(dest, path.basename(extension.projectPath));
 
 		// check if we need to write an entitlements file
-		Object.keys(extension.targetInfo).forEach(function (target) {
+		async.eachSeries(Object.keys(extension.targetInfo), function (target, next) {
 			if (!extension.targetInfo[target].entitlementsFile) {
-				return;
+				return next();
 			}
 
 			var plist = new appc.plist(fs.existsSync(extension.targetInfo[target].entitlementsFile) ? extension.targetInfo[target].entitlementsFile : null);
-			this._embedCapabilitiesAndWriteEntitlementsPlist(plist, extension.targetInfo[target].entitlementsFile);
-		}, this);
-	}, this);
+			this._embedCapabilitiesAndWriteEntitlementsPlist(plist, extension.targetInfo[target].entitlementsFile, true, next);
+		}.bind(this), next);
+	}.bind(this), next);
 };
 
 iOSBuilder.prototype.cleanXcodeDerivedData = function cleanXcodeDerivedData(next) {
@@ -6062,7 +6106,9 @@ iOSBuilder.prototype.copyResources = function copyResources(next) {
 				delete this.buildDirFiles[info.dest];
 				
 				this.cli.createHook('build.ios.copyResource', this, function(from, to, cb) {
-					if (this.copyFileSync(info.src, info.dest, { 
+					if (!fileChanged) {
+						this.logger.trace(__('No change, skipping %s', info.dest.cyan));
+					} else if (this.copyFileSync(info.src, info.dest, { 
 							//contents: contents || (contents = fs.readFileSync(info.src)), 
 							forceCopy: unsymlinkable })) {
 					if (this.useAppThinning && info.isImage && !this.forceRebuild) {
